+++ server/run.py
@@ -1,5 +1,6 @@
-from app.app import app,socketio
+from app.app import app
 from app.db import db
+from sockets.connection_socket import socketio
 
 if __name__ == "__main__" :
     db.init_app(app)
